feat: extract snippet priority input into common component
